+++ src/mainwindow.h
@@ -468,6 +468,8 @@ class MainWindow : public QMainWindow {
 
     void on_actionInsert_code_block_triggered();
 
+    void on_actionInsert_footnote_triggered();
+
     void on_actionInsert_checkbox_list_item_triggered();
 
     void on_actionNext_note_triggered();